Deduplicate github download code for rust and typescript
